+++ demisto_sdk/common/hook_validations/structure.py
@@ -42,7 +42,7 @@ class StructureValidator:
         self.scheme_name = predefined_scheme or self.scheme_of_file_by_path()
         self.file_type = self.get_file_type()
         self.current_file = self.load_data_from_file()
-        self.old_file = get_remote_file(old_file_path) if old_file_path else None
+        self.old_file = get_remote_file(old_file_path) if old_file_path else get_remote_file(file_path)
         self.configuration = configuration
 
     def is_valid_file(self):