+++ demisto_sdk/common/hook_validations/integration.py
@@ -4,6 +4,7 @@ from demisto_sdk.common.hook_validations.image import ImageValidator
 from demisto_sdk.common.hook_validations.description import DescriptionValidator
 from demisto_sdk.common.hook_validations.base_validator import BaseValidator
 from demisto_sdk.common.tools import print_error, print_warning, get_dockerimage45, server_version_compare
+from demisto_sdk.common.hook_validations.docker import DockerImageValidator
 
 
 class IntegrationValidator(BaseValidator):
@@ -56,7 +57,8 @@ class IntegrationValidator(BaseValidator):
             self.is_valid_category(),
             self.is_id_equals_name(),
             image_validator.is_valid(),
-            description_validator.is_valid()
+            description_validator.is_valid(),
+            self.is_docker_image_valid()
         ]
         return all(answers)
 
@@ -481,4 +483,12 @@ class IntegrationValidator(BaseValidator):
                 self.is_valid = False
                 return True
 
+        return False
+
+    def is_docker_image_valid(self):
+        # type: () -> bool
+        docker_image_validator = DockerImageValidator(self.file_path, is_modified_file=True, is_integration=True)
+        if not docker_image_validator.is_docker_image_valid():
+            return True
+        self.is_valid = False
         return False