+++ demisto_sdk/common/hook_validations/structure.py
@@ -37,7 +37,7 @@ class StructureValidator:
 
     def __init__(self, file_path, old_file_path=None, predefined_scheme=None, configuration=Configuration()):
         # type: (Any, Optional[str], Optional[str], Configuration) -> None
-        self.is_valid = None  # type: Optional[bool]
+        self.is_valid = True
         self.file_path = file_path  # type: ignore
         if isinstance(file_path, tuple):
             self.file_path, old_file_path = file_path[1], file_path[0]
@@ -45,7 +45,7 @@ class StructureValidator:
         self.scheme_name = predefined_scheme or self.scheme_of_file_by_path()
         self.file_type = self.get_file_type()
         self.current_file = self.load_data_from_file()
-        self.old_file = get_remote_file(old_file_path) if old_file_path else None
+        self.old_file = get_remote_file(old_file_path) if old_file_path else get_remote_file(self.file_path)
         self.configuration = configuration
 
     def is_valid_file(self):