+++ demisto_sdk/common/constants.py
@@ -6,6 +6,11 @@ class Errors:
     BACKWARDS = "Possible backwards compatibility break"
 
     @staticmethod
+    def feed_wrong_from_version(file_path, given_fromversion, needed_from_version="5.5.0"):
+        return "{} is a feed and has wrong fromversion. got `{}` expected `{}`"\
+            .format(file_path, given_fromversion, needed_from_version)
+
+    @staticmethod
     def not_used_display_name(file_path, field_name):
         return "The display details for {} will not be used in the file {} due to the type of the parameter".format(
             field_name, file_path)