+++ CHANGELOG.md
@@ -12,7 +12,9 @@
 * Added validation for docker image inside integrations and scripts.
 * Added --use-git flag to **format** command to format all changed files.
 * Fixed an issue where **validate** did not fail on dockerimage changes with bc check.
-* Fixed an issue where *unify* created a malformed integration yml.  
+* Added new flag **--ignore-entropy** to demisto-sdk **secrets**, this will allow skip entropy secrets check.
+* Added --circle flag to **lint**.
+* Fixed an issue where *unify* created a malformed integration yml.
 
 
 ### 0.3.3