+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Added validation for docker image inside integrations and scripts.
 * Added --use-git flag to **format** command to format all changed files.
 * Fixed an issue where **validate** did not fail on dockerimage changes with bc check.
+* Added new flag **--ignore-entropy** to demisto-sdk **secrets**, this will allow skip entropy secrets check.
 * Added --outfile to **lint** to allow saving failed packages to a file.