+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 * Added --use-git flag to **format** command to format all changed files.
 * Fixed an issue where **validate** did not fail on dockerimage changes with bc check.
 * Added new flag **--ignore-entropy** to demisto-sdk **secrets**, this will allow skip entropy secrets check.
-* Added --circle flag to **lint**.
+* Added --outfile to **lint** to allow saving failed packages to a file.
 
 
 ### 0.3.3